+++ packages/api/package.json
@@ -40,7 +40,6 @@
   "license": "MIT",
   "dependencies": {
     "@verdaccio/auth": "workspace:6.0.0-6-next.12",
-    "@verdaccio/commons-api": "workspace:11.0.0-6-next.4",
     "@verdaccio/config": "workspace:6.0.0-6-next.8",
     "@verdaccio/core": "workspace:6.0.0-6-next.1",
     "@verdaccio/hooks": "workspace:6.0.0-6-next.7",
