+++ packages/core/core/src/error-utils.ts
@@ -1,48 +1,5 @@
 import createError, { HttpError } from 'http-errors';
-import httpCodes from 'http-status-codes';
-import { DEFAULT_MIN_LIMIT_PASSWORD } from './constants';
-
-export const HTTP_STATUS = {
-  OK: httpCodes.OK,
-  CREATED: httpCodes.CREATED,
-  MULTIPLE_CHOICES: httpCodes.MULTIPLE_CHOICES,
-  NOT_MODIFIED: httpCodes.NOT_MODIFIED,
-  BAD_REQUEST: httpCodes.BAD_REQUEST,
-  UNAUTHORIZED: httpCodes.UNAUTHORIZED,
-  FORBIDDEN: httpCodes.FORBIDDEN,
-  NOT_FOUND: httpCodes.NOT_FOUND,
-  CONFLICT: httpCodes.CONFLICT,
-  NOT_IMPLEMENTED: httpCodes.NOT_IMPLEMENTED,
-  UNSUPPORTED_MEDIA: httpCodes.UNSUPPORTED_MEDIA_TYPE,
-  BAD_DATA: httpCodes.UNPROCESSABLE_ENTITY,
-  INTERNAL_ERROR: httpCodes.INTERNAL_SERVER_ERROR,
-  SERVICE_UNAVAILABLE: httpCodes.SERVICE_UNAVAILABLE,
-  LOOP_DETECTED: 508,
-};
-
-export const ERROR_CODE = {
-  token_required: 'token is required',
-};
-
-export const API_MESSAGE = {
-  PKG_CREATED: 'created new package',
-  PKG_CHANGED: 'package changed',
-  PKG_REMOVED: 'package removed',
-  PKG_PUBLISHED: 'package published',
-  TARBALL_UPLOADED: 'tarball uploaded successfully',
-  TARBALL_REMOVED: 'tarball removed',
-  TAG_UPDATED: 'tags updated',
-  TAG_REMOVED: 'tag removed',
-  TAG_ADDED: 'package tagged',
-  LOGGED_OUT: 'Logged out',
-};
-
-export const SUPPORT_ERRORS = {
-  PLUGIN_MISSING_INTERFACE: 'the plugin does not provide implementation of the requested feature',
-  TFA_DISABLED: 'the two-factor authentication is not yet supported',
-  STORAGE_NOT_IMPLEMENT: 'the storage does not support token saving',
-  PARAMETERS_NOT_VALID: 'the parameters are not valid',
-};
+import { DEFAULT_MIN_LIMIT_PASSWORD, HTTP_STATUS } from './constants';
 
 export const API_ERROR = {
   PASSWORD_SHORT: (passLength = DEFAULT_MIN_LIMIT_PASSWORD): string =>
@@ -80,6 +37,13 @@ export const API_ERROR = {
   BAD_PACKAGE_DATA: 'bad incoming package data',
   USERNAME_PASSWORD_REQUIRED: 'username and password is required',
   USERNAME_ALREADY_REGISTERED: 'username is already registered',
+};
+
+export const SUPPORT_ERRORS = {
+  PLUGIN_MISSING_INTERFACE: 'the plugin does not provide implementation of the requested feature',
+  TFA_DISABLED: 'the two-factor authentication is not yet supported',
+  STORAGE_NOT_IMPLEMENT: 'the storage does not support token saving',
+  PARAMETERS_NOT_VALID: 'the parameters are not valid',
 };
 
 export const APP_ERROR = {
@@ -137,8 +101,3 @@ export function getNotFound(customMessage?: string): VerdaccioError {
 export function getCode(statusCode: number, customMessage: string): VerdaccioError {
   return getError(statusCode, customMessage);
 }
-
-export const LOG_STATUS_MESSAGE =
-  "@{status}, user: @{user}(@{remoteIP}), req: '@{request.method} @{request.url}'";
-export const LOG_VERDACCIO_ERROR = `${LOG_STATUS_MESSAGE}, error: @{!error}`;
-export const LOG_VERDACCIO_BYTES = `${LOG_STATUS_MESSAGE}, bytes: @{bytes.in}/@{bytes.out}`;
